Пострался избавится от проблем. Надеюсь получилось)

+++ src/ru/fizteh/fivt/students/ermolenko/filemap/Remove.java
@@ -1,9 +1,9 @@
 package ru.fizteh.fivt.students.ermolenko.filemap;
 
-import java.io.IOException;
-
 import ru.fizteh.fivt.students.ermolenko.shell.Command;
 import ru.fizteh.fivt.students.ermolenko.shell.Shell;
+
+import java.io.IOException;
 
 public class Remove implements Command {
 
@@ -12,6 +12,10 @@ public class Remove implements Command {
     }
 
     public void executeCmd(Shell filemap, String[] args) throws IOException {
+        if (args.length != 1) {
+            System.out.println("incorrect number of arguments");
+            return;
+        }
         String key = args[0];
         String value = ((FileMap) filemap).getFileMapState().getDataBase().remove(key);
         if (value == null) {
